refactor: Moves configuration of EventSerializer to the EventStore.Configuration namespace. Removes code smell.
