+++ tests/Tutor.Web.Tests/Integration/TutorTestFactory.cs
@@ -10,7 +10,6 @@ using Tutor.Infrastructure;
 using Tutor.Infrastructure.Database;
 using Tutor.Infrastructure.Database.EventStore;
 using Tutor.Infrastructure.Database.EventStore.Configuration;
-using Tutor.Infrastructure.Database.EventStore.Configuration.EventSerializerConfiguration;
 using Tutor.Infrastructure.Database.EventStore.Postgres.Configuration;
 using Tutor.Infrastructure.Security;
